+++ examples/alpha.rs
@@ -18,7 +18,6 @@ struct MyRoguelike {
 }
 
 impl Engine for MyRoguelike {
-    fn init(&mut self, _api: &mut dyn DoryenApi) {}
     fn update(&mut self, api: &mut dyn DoryenApi) -> Option<UpdateEvent> {
         let con = api.con();
         // update the circle radius and center position
@@ -45,7 +44,6 @@ impl Engine for MyRoguelike {
             con.back(x as i32, y as i32, (0, 0, 255, 255));
         }
     }
-    fn resize(&mut self, _api: &mut dyn DoryenApi) {}
 }
 
 fn main() {