+++ examples/resize.rs
@@ -1,6 +1,6 @@
 extern crate doryen_rs;
 
-use doryen_rs::{App, AppOptions, DoryenApi, Engine, TextAlign, UpdateEvent};
+use doryen_rs::{App, AppOptions, DoryenApi, Engine, TextAlign};
 
 struct MyRoguelike {
     width: u32,
@@ -8,10 +8,6 @@ struct MyRoguelike {
 }
 
 impl Engine for MyRoguelike {
-    fn init(&mut self, _api: &mut dyn DoryenApi) {}
-    fn update(&mut self, _api: &mut dyn DoryenApi) -> Option<UpdateEvent> {
-        None
-    }
     fn render(&mut self, api: &mut dyn DoryenApi) {
         let con = api.con();
         con.rectangle(