+++ examples/subcell.rs
@@ -1,16 +1,12 @@
 extern crate doryen_rs;
 
-use doryen_rs::{App, AppOptions, DoryenApi, Engine, Image, TextAlign, UpdateEvent};
+use doryen_rs::{App, AppOptions, DoryenApi, Engine, Image, TextAlign};
 
 struct MyRoguelike {
     skull: Image,
 }
 
 impl Engine for MyRoguelike {
-    fn init(&mut self, _api: &mut dyn DoryenApi) {}
-    fn update(&mut self, _api: &mut dyn DoryenApi) -> Option<UpdateEvent> {
-        None
-    }
     fn render(&mut self, api: &mut dyn DoryenApi) {
         let con = api.con();
         con.clear(None, Some((0, 0, 0, 255)), None);
@@ -24,7 +20,6 @@ impl Engine for MyRoguelike {
             None,
         );
     }
-    fn resize(&mut self, _api: &mut dyn DoryenApi) {}
 }
 
 impl MyRoguelike {